# Faster capture-free substitution