--- a/clash-ghc/clash-ghc.cabal
+++ b/clash-ghc/clash-ghc.cabal
@@ -112,7 +112,6 @@ library
                       mtl                       >= 2.1.1    && < 2.3,
                       text                      >= 0.11.3.1 && < 1.3,
                       transformers              >= 0.4.2    && < 0.6,
-                      unbound-generics          >= 0.1      && < 0.4,
                       unordered-containers      >= 0.2.1.0  && < 0.3,
 
                       clash-lib                 >= 0.7.1    && < 1.0,