--- a/clash-ghc/src-ghc/Clash/GHC/GenerateBindings.hs
+++ b/clash-ghc/src-ghc/Clash/GHC/GenerateBindings.hs
@@ -5,23 +5,22 @@
   Maintainer  :  Christiaan Baaij <christiaan.baaij@gmail.com>
 -}
 
-{-# LANGUAGE LambdaCase      #-}
-{-# LANGUAGE TemplateHaskell #-}
-{-# LANGUAGE ViewPatterns    #-}
+{-# LANGUAGE LambdaCase        #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TemplateHaskell   #-}
+{-# LANGUAGE ViewPatterns      #-}
 
 module Clash.GHC.GenerateBindings
   (generateBindings)
 where
 
-import           Control.Lens            ((%~),(&))
+import           Control.Lens            ((%~),(&),view,_1)
 import           Control.Monad.State     (State)
 import qualified Control.Monad.State     as State
+import           Data.Coerce             (coerce)
 import           Data.Either             (lefts, rights)
-import           Data.HashMap.Strict     (HashMap)
-import qualified Data.HashMap.Strict     as HashMap
 import           Data.IntMap.Strict      (IntMap)
-import qualified Data.IntMap.Strict      as IM
-import           Unbound.Generics.LocallyNameless (bind,embed,rec,runFreshM,unembed)
+import qualified Data.IntMap.Strict      as IMS
 
 import qualified BasicTypes              as GHC
 import qualified CoreSyn                 as GHC
@@ -37,13 +36,14 @@ import           Clash.Annotations.BitRepresentation.Internal (DataRepr')
 import           Clash.Annotations.TopEntity (TopEntity)
 import           Clash.Annotations.Primitive (HDL)
 
-import           Clash.Core.Name         (Name (..), string2SystemName)
-import           Clash.Core.Term         (Term (..), TmName, TmOccName)
+import           Clash.Core.Term         (Term (..))
 import           Clash.Core.Type         (Type (..), TypeView (..), mkFunTy, splitFunForallTy, tyView)
-import           Clash.Core.TyCon        (TyCon, TyConName, TyConOccName)
+import           Clash.Core.TyCon        (TyConMap, TyConName)
 import           Clash.Core.TysPrim      (tysPrimMap)
 import           Clash.Core.Util         (mkLams, mkTyLams)
-import           Clash.Core.Var          (Var (..))
+import           Clash.Core.Var          (Var (..), Id)
+import           Clash.Core.VarEnv
+  (InScopeSet, VarEnv, extendInScopeSet, mkInScopeSet, mkVarEnv, unionVarEnv)
 import           Clash.Driver.Types      (BindingMap)
 import           Clash.GHC.GHC2Core      (GHC2CoreState, tyConMap, coreToId, coreToName, coreToTerm,
                                           makeAllTyCons, qualfiedNameString, emptyGHC2CoreState)
@@ -51,6 +51,8 @@ import           Clash.GHC.LoadModules   (loadModules)
 import           Clash.Primitives.Types  (PrimMap, ResolvedPrimMap, Primitive)
 import           Clash.Primitives.Util   (generatePrimMap)
 import           Clash.Rewrite.Util      (mkInternalVar, mkSelectorCase)
+import           Clash.Unique
+  (listToUniqMap, lookupUniqMap, mapUniqMap, unionUniqMap, uniqMapToUniqSet)
 import           Clash.Util              ((***),first)
 
 generateBindings
@@ -63,12 +65,11 @@ generateBindings
   -> String
   -> Maybe  (GHC.DynFlags)
   -> IO ( BindingMap
-        , HashMap TyConOccName TyCon
+        , TyConMap
         , IntMap TyConName
-        , [( TmName          -- topEntity bndr
-           , Type            -- type of the topEntity bndr
+        , [( Id
            , Maybe TopEntity -- (maybe) TopEntity annotation
-           , Maybe TmName    -- (maybe) associated testbench
+           , Maybe Id        -- (maybe) associated testbench
            )]
         , ResolvedPrimMap  -- The primitives found in '.' and 'primDir'
         , [DataRepr']
@@ -79,17 +80,20 @@ generateBindings primDirs importDirs hdl modName dflagsM = do
   let ((bindingsMap,clsVMap),tcMap) = State.runState (mkBindings primMap bindings clsOps unlocatable) emptyGHC2CoreState
       (tcMap',tupTcCache)           = mkTupTyCons tcMap
       tcCache                       = makeAllTyCons tcMap' fiEnvs
-      allTcCache                    = tysPrimMap `HashMap.union` tcCache
-      clsMap                        = HashMap.map (\(nm,ty,i) -> (nm,ty,GHC.noSrcSpan,GHC.Inline,mkClassSelector allTcCache ty i)) clsVMap
-      allBindings                   = bindingsMap `HashMap.union` clsMap
+      allTcCache                    = tysPrimMap `unionUniqMap` tcCache
+      inScope0 = mkInScopeSet (uniqMapToUniqSet
+                      ((mapUniqMap (coerce . view _1) bindingsMap) `unionUniqMap`
+                       (mapUniqMap (coerce . view _1) clsMap)))
+      clsMap                        = mapUniqMap (\(v,i) -> (v,GHC.noSrcSpan,GHC.Inline,mkClassSelector inScope0 allTcCache (varType v) i)) clsVMap
+      allBindings                   = bindingsMap `unionVarEnv` clsMap
       topEntities'                  =
         flip State.evalState tcMap' $ mapM (\(topEnt,annM,benchM) -> do
           topEnt' <- coreToName GHC.varName GHC.varUnique qualfiedNameString topEnt
-          benchM' <- traverse (coreToName GHC.varName GHC.varUnique qualfiedNameString) benchM
+          benchM' <- traverse coreToId benchM
           return (topEnt',annM,benchM')) topEntities
-      topEntities''                 = map (\(topEnt,annM,benchM) -> case HashMap.lookup (nameOcc topEnt) allBindings of
-                                              Just (_,ty,_,_,_) -> (topEnt,ty,annM,benchM)
-                                              Nothing       -> error "This shouldn't happen"
+      topEntities''                 = map (\(topEnt,annM,benchM) -> case lookupUniqMap topEnt allBindings of
+                                              Just (v,_,_,_) -> (v,annM,benchM)
+                                              Nothing        -> error "This shouldn't happen"
                                           ) topEntities'
 
   return (allBindings, allTcCache, tupTcCache, topEntities'', primMap, reprs)
@@ -104,7 +108,7 @@ mkBindings
   -- Unlocatable Expressions
   -> State GHC2CoreState
            ( BindingMap
-           , HashMap TmOccName (TmName,Type,Int)
+           , VarEnv (Id,Int)
            )
 mkBindings primMap bindings clsOps unlocatable = do
   bindingsList <- mapM (\case
@@ -113,7 +117,7 @@ mkBindings primMap bindings clsOps unlocatable = do
                                 inl = GHC.inlinePragmaSpec . GHC.inlinePragInfo $ GHC.idInfo v
                             tm <- coreToTerm primMap unlocatable sp e
                             v' <- coreToId v
-                            return [(nameOcc (varName v'), (varName v',unembed (varType v'), sp, inl, tm))]
+                            return [(v', (v', sp, inl, tm))]
                           GHC.Rec bs -> do
                             tms <- mapM (\(v,e) -> do
                                           let sp = GHC.getSrcSpan v
@@ -122,47 +126,46 @@ mkBindings primMap bindings clsOps unlocatable = do
                                           return (v',sp,tm)
                                         ) bs
                             case tms of
-                              [(v,sp,tm)] -> return [(nameOcc (varName v), (varName v,unembed (varType v), sp, GHC.NoInline, tm))]
-                              _ ->
-                                return $ map (\(v,sp,e) -> (nameOcc (varName v),(varName v,unembed (varType v),sp,GHC.NoInline
-                                                  ,Letrec (bind (rec (map (\(x,_,y) -> (x,embed y)) tms)) e)))) tms
+                              [(v,sp,tm)] -> return [(v, (v, sp, GHC.NoInline, tm))]
+                              _ -> return $ map (\(v,sp,e) -> (v,(v,sp,GHC.NoInline, Letrec (map (\(x,_,y) -> (x,y)) tms) e))) tms
                        ) bindings
   clsOpList    <- mapM (\(v,i) -> do
                           v' <- coreToId v
-                          let ty = unembed $ varType v'
-                          return (nameOcc (varName v'), (varName v',ty,i))
+                          return (v', (v',i))
                        ) clsOps
 
-  return (HashMap.fromList (concat bindingsList), HashMap.fromList clsOpList)
+  return (mkVarEnv (concat bindingsList), mkVarEnv clsOpList)
 
 mkClassSelector
-  :: HashMap TyConOccName TyCon
+  :: InScopeSet
+  -> TyConMap
   -> Type
   -> Int
   -> Term
-mkClassSelector tcm ty sel = newExpr
+mkClassSelector inScope0 tcm ty sel = newExpr
   where
     ((tvs,dictTy:_),_) = first (lefts *** rights)
                        $ first (span (\l -> case l of Left _ -> True
                                                       _      -> False))
                        $ splitFunForallTy ty
     newExpr = case tyView dictTy of
-      (TyConApp _ _) -> runFreshM $ flip State.evalStateT (0 :: Int) $ do
-                          (dcId,dcVar) <- mkInternalVar (string2SystemName "dict") dictTy
-                          selE         <- mkSelectorCase "mkClassSelector" tcm dcVar 1 sel
+      (TyConApp _ _) -> flip State.evalState (0 :: Int) $ do
+                          dcId <- mkInternalVar inScope0 "dict" dictTy
+                          let inScope1 = extendInScopeSet inScope0 dcId
+                          selE <- mkSelectorCase "mkClassSelector" inScope1 tcm (Var dcId) 1 sel
                           return (mkTyLams (mkLams selE [dcId]) tvs)
-      (FunTy arg res) -> runFreshM $ flip State.evalStateT (0 :: Int) $ do
-                           (dcId,dcVar) <- mkInternalVar (string2SystemName "dict") (mkFunTy arg res)
-                           return (mkTyLams (mkLams dcVar [dcId]) tvs)
-      (OtherType oTy) -> runFreshM $ flip State.evalStateT (0 :: Int) $ do
-                           (dcId,dcVar) <- mkInternalVar (string2SystemName "dict") oTy
-                           return (mkTyLams (mkLams dcVar [dcId]) tvs)
+      (FunTy arg res) -> flip State.evalState (0 :: Int) $ do
+                           dcId <- mkInternalVar inScope0 "dict" (mkFunTy arg res)
+                           return (mkTyLams (mkLams (Var dcId) [dcId]) tvs)
+      (OtherType oTy) -> flip State.evalState (0 :: Int) $ do
+                           dcId <- mkInternalVar inScope0 "dict" oTy
+                           return (mkTyLams (mkLams (Var dcId) [dcId]) tvs)
 
 mkTupTyCons :: GHC2CoreState -> (GHC2CoreState,IntMap TyConName)
 mkTupTyCons tcMap = (tcMap'',tupTcCache)
   where
     tupTyCons        = map (GHC.tupleTyCon GHC.Boxed) [2..62]
     (tcNames,tcMap') = State.runState (mapM (\tc -> coreToName GHC.tyConName GHC.tyConUnique qualfiedNameString tc) tupTyCons) tcMap
-    tupTcCache       = IM.fromList (zip [2..62] tcNames)
-    tupHM            = HashMap.fromList (zip (map nameOcc tcNames) tupTyCons)
-    tcMap''          = tcMap' & tyConMap %~ (`HashMap.union` tupHM)
+    tupTcCache       = IMS.fromList (zip [2..62] tcNames)
+    tupHM            = listToUniqMap (zip tcNames tupTyCons)
+    tcMap''          = tcMap' & tyConMap %~ (`unionUniqMap` tupHM)