--- a/clash-lib/src/Clash/Backend.hs
+++ b/clash-lib/src/Clash/Backend.hs
@@ -14,8 +14,9 @@ import qualified  Control.Lens              as Lens
 import Data.HashSet                         (HashSet)
 import Data.Maybe                           (fromMaybe)
 import Data.Semigroup.Monad                 (Mon (..))
-import qualified Data.Text.Lazy             as T
-import Data.Text.Lazy                       (Text)
+import qualified Data.Text                  as T
+import Data.Text                            (Text)
+import qualified Data.Text.Lazy             as LT
 import Control.Monad.State                  (State)
 import Data.Text.Prettyprint.Doc.Extra      (Doc)
 
@@ -27,7 +28,7 @@ import Clash.Netlist.BlackBox.Types
 
 import Clash.Annotations.Primitive          (HDL)
 
-type ModName = String
+type ModName = Identifier
 
 -- | Is a type used for internal or external use
 data Usage
@@ -57,9 +58,9 @@ class Backend state where
   extractTypes     :: state -> HashSet HWType
 
   -- | Generate HDL for a Netlist component
-  genHDL           :: String -> SrcSpan -> [Identifier] -> Component -> Mon (State state) ((String, Doc),[(String,Doc)])
+  genHDL           :: Identifier -> SrcSpan -> [Identifier] -> Component -> Mon (State state) ((String, Doc),[(String,Doc)])
   -- | Generate a HDL package containing type definitions for the given HWTypes
-  mkTyPackage      :: String -> [HWType] -> Mon (State state) [(String, Doc)]
+  mkTyPackage      :: Identifier -> [HWType] -> Mon (State state) [(String, Doc)]
   -- | Convert a Netlist HWType to a target HDL type
   hdlType          :: Usage -> HWType -> Mon (State state) Doc
   -- | Convert a Netlist HWType to an HDL error value for that type
@@ -69,7 +70,7 @@ class Backend state where
   -- | Create a record selector
   hdlRecSel        :: HWType -> Int -> Mon (State state) Doc
   -- | Create a signal declaration from an identifier (Text) and Netlist HWType
-  hdlSig           :: Text -> HWType -> Mon (State state) Doc
+  hdlSig           :: LT.Text -> HWType -> Mon (State state) Doc
   -- | Create a generative block statement marker
   genStmt          :: Bool -> State state Doc
   -- | Turn a Netlist Declaration to a HDL concurrent block
@@ -79,9 +80,9 @@ class Backend state where
   -- | Bit-width of Int/Word/Integer
   iwWidth          :: State state Int
   -- | Convert to a bit-vector
-  toBV             :: HWType -> Text -> Mon (State state) Doc
+  toBV             :: HWType -> LT.Text -> Mon (State state) Doc
   -- | Convert from a bit-vector
-  fromBV           :: HWType -> Text -> Mon (State state) Doc
+  fromBV           :: HWType -> LT.Text -> Mon (State state) Doc
   -- | Synthesis tool we're generating HDL for
   hdlSyn           :: State state HdlSyn
   -- | mkIdentifier
@@ -99,8 +100,8 @@ class Backend state where
   -- | unextend/unescape identifier
   unextend         :: State state (Identifier -> Identifier)
   addIncludes      :: [(String, Doc)] -> State state ()
-  addLibraries     :: [Text] -> State state ()
-  addImports       :: [Text] -> State state ()
+  addLibraries     :: [LT.Text] -> State state ()
+  addImports       :: [LT.Text] -> State state ()
   addAndSetData    :: FilePath -> State state String
   getDataFiles     :: State state [(String,FilePath)]
   addMemoryDataFile  :: (String,String) -> State state ()