--- a/clash-lib/src/Clash/Core/Subst.hs-boot
+++ b/clash-lib/src/Clash/Core/Subst.hs-boot
@@ -1,0 +1,15 @@
+module Clash.Core.Subst where
+
+import {-# SOURCE #-} Clash.Core.Type (Type)
+import Clash.Core.Var (TyVar)
+
+substTyWith
+  :: [TyVar]
+  -> [Type]
+  -> Type
+  -> Type
+
+aeqType
+  :: Type
+  -> Type
+  -> Bool