--- a/clash-lib/src/Clash/Core/Term.hs-boot
+++ b/clash-lib/src/Clash/Core/Term.hs-boot
@@ -12,6 +12,7 @@ import GHC.Generics    (Generic)
 import Clash.Core.Name (Name)
 
 data Term
+
 type TmName = Name Term
 
 instance Generic Term