… benchmark/benchmark-normalization.hs
… benchmark/common/BenchmarkCommon.hs
… benchmark/profiling/run/profile-normalization-run.hs
… clash-ghc/src-ghc/Clash/GHC/Evaluator.hs
… clash-ghc/src-ghc/Clash/GHC/GHC2Core.hs
… clash-lib/src/Clash/Backend/SystemVerilog.hs
… clash-lib/src/Clash/Backend/VHDL.hs
… clash-lib/src/Clash/Backend/Verilog.hs
… clash-lib/src/Clash/Core/DataCon.hs
… clash-lib/src/Clash/Core/Evaluator.hs
… clash-lib/src/Clash/Core/FreeVars.hs
… clash-lib/src/Clash/Core/Literal.hs
… clash-lib/src/Clash/Core/Name.hs
… clash-lib/src/Clash/Core/Pretty.hs
… clash-lib/src/Clash/Core/Subst.hs
… clash-lib/src/Clash/Core/Term.hs
… clash-lib/src/Clash/Core/TyCon.hs
… clash-lib/src/Clash/Core/TyCon.hs-boot
… clash-lib/src/Clash/Core/Type.hs
… clash-lib/src/Clash/Core/Type.hs-boot
… clash-lib/src/Clash/Core/TysPrim.hs
… clash-lib/src/Clash/Core/Util.hs
… clash-lib/src/Clash/Core/Var.hs
… clash-lib/src/Clash/Core/VarEnv.hs
… clash-lib/src/Clash/Driver.hs
… clash-lib/src/Clash/Driver/Types.hs
… clash-lib/src/Clash/Netlist.hs
… clash-lib/src/Clash/Netlist.hs-boot
… clash-lib/src/Clash/Netlist/BlackBox.hs
… clash-lib/src/Clash/Netlist/BlackBox/Util.hs