This patch removes the use of the `unbound-generics` for names
and bound-variables, as profiling indicated that at least half
of the time of the normalisation pass was spend inside functions
concerning themselves with keep invariants alive for the
substitution functions of `unbound-generics`.

In stead we implement our own substitution, alpha-equivalence,
and free variable calculation; and use the capture-free
substitution method referred to as "The rapier" in section 3.2 of

      "Secrets of the GHC inliner"

Since we change the representation for names, this patch
touches a lot of modules, which cannot be split into multiple
commits.

As a result of this patch, we get a 10x to sometimes 50x speedup
on the normalisation pass; which is the bulk of the work done
by the Clash compiler.

### Baseline

```
examples/FIR.hs
time                 831.1 ms   (824.4 ms .. 834.2 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 829.0 ms   (825.8 ms .. 830.4 ms)
std dev              2.291 ms   (785.2 μs .. 3.110 ms)
variance introduced by outliers: 19% (moderately inflated)

examples/Queens.hs
time                 151.5 ms   (146.0 ms .. 154.1 ms)
                     0.999 R²   (0.996 R² .. 1.000 R²)
mean                 151.2 ms   (148.6 ms .. 153.8 ms)
std dev              3.875 ms   (2.584 ms .. 5.868 ms)
variance introduced by outliers: 12% (moderately inflated)

benchmark/tests/BundleMapRepeat.hs
time                 4.919 s    (4.907 s .. 4.933 s)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 4.918 s    (4.914 s .. 4.921 s)
std dev              4.190 ms   (2.282 ms .. 5.910 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmark/tests/PipelinesViaFolds.hs
time                 6.672 s    (5.967 s .. 7.360 s)
                     0.999 R²   (0.995 R² .. 1.000 R²)
mean                 6.863 s    (6.666 s .. 7.143 s)
std dev              270.8 ms   (76.54 ms .. 366.9 ms)
variance introduced by outliers: 19% (moderately inflated)
``` 

### The Rapier

```
examples/FIR.hs
time                 68.83 ms   (67.80 ms .. 69.76 ms) (12x speedup over baseline)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 68.38 ms   (67.73 ms .. 69.12 ms) (12x speedup over baseline)
std dev              1.294 ms   (900.2 μs .. 1.742 ms)

examples/Queens.hs
time                 12.95 ms   (12.62 ms .. 13.24 ms) (11.5x speedup over baseline)
                     0.995 R²   (0.992 R² .. 0.998 R²)
mean                 12.77 ms   (12.61 ms .. 13.03 ms) (12x speedup over baseline)
std dev              527.8 μs   (430.0 μs .. 727.1 μs)
variance introduced by outliers: 17% (moderately inflated)

benchmark/tests/BundleMapRepeat.hs
time                 122.1 ms   (119.9 ms .. 123.4 ms) (40x speedup over baseline)
                     1.000 R²   (0.998 R² .. 1.000 R²)
mean                 120.1 ms   (118.3 ms .. 121.5 ms) (41x speedup over baseline)
std dev              2.412 ms   (1.736 ms .. 3.080 ms)
variance introduced by outliers: 11% (moderately inflated)

benchmark/tests/PipelinesViaFolds.hs
time                 128.5 ms   (125.7 ms .. 131.3 ms) (52x speedup over baseline)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 128.5 ms   (127.8 ms .. 130.2 ms) (53.5x speedup over baseline)
std dev              1.887 ms   (1.356 ms .. 2.642 ms)
variance introduced by outliers: 11% (moderately inflated)
```