@martijnbastiaan 2018-10-09
Rebased on `master`. Original diff on top of `1ee941f6525369bf6fe708d56a6513f5b42aa7ed`: [361.diff.zip](https://github.com/clash-lang/clash-compiler/files/2460515/361.diff.zip).