@martijnbastiaan 2018-10-09
Did you add the following on purpose in `cabal.project`?

```diff
+ package clash-lib
+   flags: debug
+ 
```