@christiaanb 2018-10-09
Yes, as a dev you want to run the debug version by default